app/bluetooth/example_host/bt_host_esl_ap/requirements.txt
numpy
pillow
cryptography
colorlog
intelhex
qrcode
ctypesgen
pyserial
